refac: extract `instructions::context` mod
