+++ crates/interpreter/src/instructions.rs
@@ -5,6 +5,7 @@ pub mod macros;
 pub mod arithmetic;
 pub mod bitwise;
 pub mod block_info;
+pub mod context;
 pub mod contract;
 pub mod control;
 pub mod data;
@@ -19,7 +20,7 @@ pub mod utility;
 use crate::{interpreter_types::InterpreterTypes, Host};
 
 /// EVM opcode function signature.
-pub type Instruction<W, H> = fn(&mut control::InstructionContext<'_, H, W>);
+pub type Instruction<W, H> = fn(&mut context::InstructionContext<'_, H, W>);
 
 /// Instruction table is list of instruction function pointers mapped to 256 EVM opcodes.
 pub type InstructionTable<W, H> = [Instruction<W, H>; 256];
