+++ crates/interpreter/src/instructions/block_info.rs
@@ -5,7 +5,7 @@ use crate::{
 };
 use primitives::{hardfork::SpecId::*, U256};
 
-use super::control::InstructionContext;
+use super::context::InstructionContext;
 
 /// EIP-1344: ChainID opcode
 pub fn chainid<WIRE: InterpreterTypes, H: Host + ?Sized>(
