+++ crates/interpreter/src/instructions/control.rs
@@ -9,7 +9,7 @@ use crate::{
 };
 use primitives::{Bytes, U256};
 
-use super::Instruction;
+use super::context::InstructionContext;
 
 pub fn rjump<WIRE: InterpreterTypes, H: Host + ?Sized>(
     context: &mut InstructionContext<'_, H, WIRE>,
@@ -55,35 +55,6 @@ pub fn rjumpv<WIRE: InterpreterTypes, H: Host + ?Sized>(
         offset += context.interpreter.bytecode.read_offset_i16(1 + case * 2) as isize;
     }
     context.interpreter.bytecode.relative_jump(offset);
-}
-
-pub struct InstructionContext<'a, H: Host + ?Sized, ITy: InterpreterTypes> {
-    pub host: &'a mut H,
-    pub interpreter: &'a mut Interpreter<ITy>,
-}
-
-impl<H: Host + ?Sized, ITy: InterpreterTypes> InstructionContext<'_, H, ITy> {
-    /// Executes the instruction at the current instruction pointer.
-    ///
-    /// Internally it will increment instruction pointer by one.
-    #[inline]
-    pub(crate) fn step(&mut self, instruction_table: &[Instruction<ITy, H>; 256]) {
-        // Get current opcode.
-        let opcode = self.interpreter.bytecode.opcode();
-
-        // SAFETY: In analysis we are doing padding of bytecode so that we are sure that last
-        // byte instruction is STOP so we are safe to just increment program_counter bcs on last instruction
-        // it will do noop and just stop execution of this contract
-        self.interpreter.bytecode.relative_jump(1);
-
-        // Execute instruction.
-        instruction_table[opcode as usize](self)
-    }
-
-    #[inline]
-    pub(crate) fn can_continue(&mut self) -> bool {
-        self.interpreter.control.instruction_result().is_continue()
-    }
 }
 
 pub fn jump<ITy: InterpreterTypes, H: Host + ?Sized>(context: &mut InstructionContext<'_, H, ITy>) {
