+++ crates/interpreter/src/instructions/tx_info.rs
@@ -5,7 +5,7 @@ use crate::{
 };
 use primitives::U256;
 
-use super::control::InstructionContext;
+use super::context::InstructionContext;
 
 pub fn gasprice<WIRE: InterpreterTypes, H: Host + ?Sized>(
     context: &mut InstructionContext<'_, H, WIRE>,
